actividades_cc/act_semana_2/requirements.txt
pip == 24

matplotlib
numpy == 1.23.5
pandas

scikit-learn==1.4.2
joblib

azureml-sdk